função move movida para Snake.js
