+++ index.js
@@ -1,6 +1,6 @@
 import { drawBackground, drawFruit, drawLingua, drawSnake } from "./src/View.js"
 import { getData } from "./src/Data.js"
-import { snake, setSnakeSettings, moves } from "./src/Snake.js"
+import { snake, setSnakeSettings, moves, move } from "./src/Snake.js"
 import { 
     game, setGameSettings, pause, gameOver
 } from "./src/Game.js"
@@ -16,37 +16,6 @@ const comands = {
 const randItem = arr => {
     const randint = (min,max) => Math.floor(Math.random() * (max-min+1)) + min
     return arr[randint(0,arr.length - 1)]
-}
-
-const move = async () => {
-    const { direcao } = snake
-    const { quantX, quantY } = game
-    
-    if(direcao !== null){
-        snake.cauda.push( {px: snake.px, py: snake.py})
-        snake.ultima = snake.cauda.shift()
-    }
-
-    if(direcao === "up"){
-        return snake.py = (snake.py - 1) >= 0 ? snake.py - 1 : quantY - 1
-    }
-    if(direcao === "left"){
-        return snake.px = (snake.px - 1) >= 0 ? snake.px - 1 : quantX - 1
-    }
-
-    if(direcao === "down"){
-        if(snake.py + 1 < quantY){
-            return snake.py += 1
-        }
-        return snake.py = snake.py - quantY + 1
-    }
-
-    if(direcao === "right"){
-        if(snake.px + 1 < quantX){
-            return snake.px += 1
-        }
-        return snake.px = snake.px - quantX + 1
-    }
 }
 
 const spawFruit = async () => {
