+++ src/Snake.js
@@ -1,4 +1,5 @@
 import { getDefaultSnake } from "./Settings.js"
+import { game } from "./Game.js"
 
 const snake = {
     setDirecao(direcao){
@@ -28,12 +29,6 @@ const snake = {
     }
 }
 
-const setSnakeSettings = async ()=> {
-    let settings = await getDefaultSnake()
-    settings = Object.entries(settings)
-    settings.forEach( s => snake[s[0]] = s[1])
-}
-
 const moves = {
     "ArrowLeft": () => snake.setDirecao("left"),
     "ArrowDown": () => snake.setDirecao("down"),
@@ -45,4 +40,41 @@ const moves = {
     "d": () => snake.setDirecao("right")
 }
 
-export { snake, setSnakeSettings, moves }
+const setSnakeSettings = async ()=> {
+    let settings = await getDefaultSnake()
+    settings = Object.entries(settings)
+    settings.forEach( s => snake[s[0]] = s[1])
+}
+
+const move = async () => {
+    const { direcao } = snake
+    const { quantX, quantY } = game
+    
+    if(direcao !== null){
+        snake.cauda.push( {px: snake.px, py: snake.py})
+        snake.ultima = snake.cauda.shift()
+    }
+
+    if(direcao === "up"){
+        return snake.py = (snake.py - 1) >= 0 ? snake.py - 1 : quantY - 1
+    }
+    if(direcao === "left"){
+        return snake.px = (snake.px - 1) >= 0 ? snake.px - 1 : quantX - 1
+    }
+
+    if(direcao === "down"){
+        if(snake.py + 1 < quantY){
+            return snake.py += 1
+        }
+        return snake.py = snake.py - quantY + 1
+    }
+
+    if(direcao === "right"){
+        if(snake.px + 1 < quantX){
+            return snake.px += 1
+        }
+        return snake.px = snake.px - quantX + 1
+    }
+}
+
+export { snake, setSnakeSettings, moves, move }
